Moving code from Geopoint codec to Geopoint class
