+++ packages/xforms-engine/src/lib/codecs/Geopoint/GeopointValueCodec.ts
@@ -1,22 +1,5 @@
 import { type CodecDecoder, type CodecEncoder, ValueCodec } from '../ValueCodec.ts';
-import { Geopoint, type GeopointRuntimeValue } from './Geopoint.ts';
-
-// TODO: Add support for GeoJSONValue
-export type GeopointInputValue = GeopointRuntimeValue | string;
-
-const decodeStringValue = (value: GeopointInputValue): GeopointRuntimeValue => {
-	if (typeof value !== 'string' || value.trim() === '') {
-		return null;
-	}
-
-	const [latitude, longitude, altitude = null, accuracy = null] = value.split(/\s+/).map(Number);
-
-	if (latitude == null || longitude == null || Number.isNaN(altitude) || Number.isNaN(accuracy)) {
-		return null;
-	}
-
-	return new Geopoint({ latitude, longitude, altitude, accuracy }).getRuntimeValue();
-};
+import { Geopoint, type GeopointInputValue, type GeopointRuntimeValue } from './Geopoint.ts';
 
 export class GeopointValueCodec extends ValueCodec<
 	'geopoint',
@@ -25,20 +8,11 @@ export class GeopointValueCodec extends ValueCodec<
 > {
 	constructor() {
 		const encodeValue: CodecEncoder<GeopointInputValue> = (value) => {
-			const geopointValue = typeof value === 'string' ? decodeStringValue(value) : value;
-
-			if (geopointValue == null) {
-				return '';
-			}
-
-			return new Geopoint(geopointValue)
-				.getTuple()
-				.map((item) => item.value ?? 0)
-				.join(' ');
+			return Geopoint.toCoordinatesString(value);
 		};
 
 		const decodeValue: CodecDecoder<GeopointRuntimeValue> = (value: string) => {
-			return decodeStringValue(value);
+			return Geopoint.parseString(value);
 		};
 
 		super('geopoint', encodeValue, decodeValue);
